feat(servers): add state list tab and improve data page layout
- Add new state list tab component to display server state data
- Move page.client.tsx logic into page.tsx and remove client component
- Update icon sizes in server layout for consistency
- Make error messages bold for better visibility
- Reorganize imports and fix minor code formatting issues
