+++ app/[locale]/(main)/servers/[id]/data/build-destroy-list-footer.tsx
@@ -1,6 +1,8 @@
-import PaginationNavigator from '@/components/common/pagination-navigator';
-import usePathId from '@/hooks/use-path-id';
+'use client';
 
+import PaginationNavigator from '@/components/common/pagination-navigator';
+
+import usePathId from '@/hooks/use-path-id';
 
 export default function BuildDestroyListFooter() {
 	const id = usePathId();
