+++ app/[locale]/(main)/servers/[id]/layout.tsx
@@ -3,6 +3,7 @@
 import {
 	BarChart4,
 	FileIcon,
+	HardDriveIcon,
 	HistoryIcon,
 	LayoutDashboardIcon,
 	MapIcon,
@@ -92,49 +93,49 @@ export default function ServerLayout({ params, children }: LayoutProps) {
 				id: 'console',
 				href: '/console',
 				label: <Tran text="console" />,
-				icon: <TerminalIcon />,
+				icon: <TerminalIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'map', //
 				href: '/maps',
 				label: <Tran text="map" />,
-				icon: <MapIcon />,
+				icon: <MapIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'plugin',
 				href: '/plugins',
 				label: <PluginLabel />,
-				icon: <PlugIcon />,
+				icon: <PlugIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'data',
 				href: '/data',
 				label: <Tran text="data" />,
-				icon: <HistoryIcon />,
+				icon: <HardDriveIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'metric',
 				href: '/metrics',
 				label: <Tran text="metric" />,
-				icon: <BarChart4 />,
+				icon: <BarChart4 className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'file', //
 				href: '/files',
 				label: <Tran text="file" />,
-				icon: <FileIcon />,
+				icon: <FileIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 			{
 				id: 'setting', //
 				href: '/setting',
 				label: <Tran text="setting" />,
-				icon: <SettingsIcon />,
+				icon: <SettingsIcon className="size-5" />,
 				filter: { any: [{ authority: 'UPDATE_SERVER' }, { authorId: ownerId }] },
 			},
 		],
