+++ components/common/error-message.tsx
@@ -13,8 +13,8 @@ export default function ErrorMessage({ className, error }: Props) {
 	const message = getErrorMessage(error);
 
 	if (message) {
-		return <span className={cn('text-destructive-foreground p-2 text-sm', className)}>{message}</span>;
+		return <span className={cn('text-destructive-foreground font-semibold p-2 text-sm', className)}>{message}</span>;
 	}
 
-	return <Tran className={cn('text-destructive-foreground p-2 text-sm', className)} text="unknown-error" />;
+	return <Tran className={cn('text-destructive-foreground font-semibold p-2 text-sm', className)} text="unknown-error" />;
 }
